Seperate into classes and put in package sisyphus
