+++ src/main/java/sisyphus/TaskList.java
@@ -1,0 +1,35 @@
+package sisyphus;
+
+import java.util.ArrayList;
+
+public class TaskList {
+    private ArrayList<Task> tasks;
+
+    public TaskList(ArrayList<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public void addTask(Task task) {
+        tasks.add(task);
+    }
+
+    public void removeTask(int number) {
+        tasks.remove(number - 1);
+    }
+
+    public boolean isEmpty() {
+        return tasks.isEmpty();
+    }
+
+    public int size() {
+        return tasks.size();
+    }
+
+    public Task get(int index) {
+        return tasks.get(index);
+    }
+
+    public ArrayList<Task> getTasks() {
+        return tasks;
+    }
+}
